[BE] [feat] : 답례품 통계 기능 작성

+++ backend/givutake/src/main/java/com/accepted/givutake/funding/repository/FundingStatisticsRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 
 @Repository
-public interface FundingStatisticsRepository extends JpaRepository<FundingParticipants, Long> {
+public interface FundingStatisticsRepository extends JpaRepository<FundingParticipants, Integer> {
     @Query(nativeQuery = true, value = """
             SELECT
                 IF(u.is_male = TRUE, 'male', 'female') AS gender,
